MAM-56: Merge common functionality and get Join voting finished state working
