+++ mamba/mamba/Planning/Join/Session Landing/ViewModels/PlanningJoinSessionLandingViewModel.swift
@@ -9,34 +9,18 @@
 import Foundation
 import Combine
 
-class PlanningJoinSessionLandingViewModel: ObservableObject {
-    private var service: PlanningJoinSessionLandingServiceProtocol
-    private var sessionCode: String
-    private var participantName: String
+class PlanningJoinSessionLandingViewModel: PlanningSessionLandingViewModel<PlanningCommands.JoinSend, PlanningCommands.JoinReceive> {
     private var cancellable: AnyCancellable?
-    private(set) var availableCards: [PlanningCard] = []
-    @Published var state: PlanningSessionLandingState = .loading
-    @Published var sessionName: String = ""
-    @Published var participants = [PlanningParticipant]()
-    @Published var ticket: PlanningTicket?
-    @Published var selectedCard: PlanningCard? {
-        didSet {
-            sendVoteCommand()
-        }
-    }
-    
-    var participantList: [PlanningParticipantRowViewModel] {
-        participants.map {
-            PlanningParticipantRowViewModel(participantName: $0.name,
-                                            votingValue: participantVotedValue($0) ?? "")
-        }
-    }
     
     init(sessionCode: String, participantName: String) {
-        self.service = PlanningJoinSessionLandingService()
-        self.sessionCode = sessionCode
-        self.participantName = participantName
-        startSession()
+        super.init(websocketURL: URLCenter.shared.planningJoinWSURL)
+        configure(sessionCode: sessionCode)
+        configure(participantName: participantName)
+        
+        cancellable = $selectedCard.sink { selectedCard in
+            guard let card = selectedCard else { return }
+            self.sendVoteCommand(card)
+        }
     }
     
     func sendJoinSessionCommand() {
@@ -44,68 +28,14 @@ class PlanningJoinSessionLandingViewModel: ObservableObject {
         sendCommand(.joinSession(commandMessage))
     }
     
-    private func participantVotedValue(_ participant: PlanningParticipant) -> String? {
-        switch state {
-        case .finishedVoting:
-            let selectedCard = ticket?.ticketVotes.first(where: { $0.user.id == participant.id })?.selectedCard
-            return selectedCard?.title ?? "Skipped"
-        default:
-            return nil
-        }
-    }
-    
-    private func sendVoteCommand() {
-        guard
-            let ticketId = ticket?.identifier,
-            let selectedCard = selectedCard
-        else {
-            return
-        }
+    private func sendVoteCommand(_ selectedCard: PlanningCard) {
+        guard let ticketId = ticket?.identifier else { return }
         let commandMessage = PlanningVoteMessage(ticketId: ticketId, selectedCard: selectedCard)
         sendCommand(.vote(commandMessage))
     }
     
-    private func sendCommand(_ command: PlanningCommands.JoinSend) {
-        Log.log(level: .debug, category: .planning, message: "Join sending command: %@", args: String(describing: command))
-        do {
-            try service.sendCommand(command)
-        } catch let error as EncodingError {
-            executeError(code: error.errorCode, description: error.errorCustomDescription)
-        } catch {
-            executeError(code: "3105", description: error.localizedDescription)
-        }
-    }
-    
-    private func startSession() {
-        guard cancellable == nil else { return }
-        cancellable = service.startSession()
-            .receive(on: DispatchQueue.main)
-            .sink(receiveCompletion: { networkError in
-                switch networkError {
-                case .finished:
-                    //TODO: MAM-61
-                    break
-                case .failure(let error):
-                    self.executeError(code: error.errorCode, description: error.errorDescription)
-                }
-            }, receiveValue: { result in
-                switch result {
-                case .success(let command):
-                    self.executeCommand(command)
-                case .failure(let error):
-                    self.executeError(code: error.errorCode, description: error.errorDescription)
-                }
-            })
-    }
-    
-    private func executeError(code: String, description: String) {
-        Log.log(level: .error, category: .planning, message: "Join executing error state: %@-%@", args: code, description)
-        let planningError = PlanningLandingError(code: code, description: description)
-        self.state = .error(planningError)
-    }
-    
-    private func executeCommand(_ command: PlanningCommands.JoinReceive) {
-        Log.log(level: .debug, category: .planning, message: "Join executing received command: %{private}@", args: String(describing: command))
+    public override func executeCommand(_ command: PlanningCommands.JoinReceive) {
+        super.executeCommand(command)
         switch command {
         case .noneState(let message):
             state = .none
@@ -128,16 +58,5 @@ class PlanningJoinSessionLandingViewModel: ObservableObject {
             break
         }
     }
-    
-    private func parseStateMessage(_ message: PlanningSessionStateMessage) {
-        self.participants = message.participants
-        self.sessionCode = message.sessionCode
-        self.sessionName = message.sessionName
-        self.ticket = message.ticket
-        self.availableCards = message.availableCards
-        
-        for participant in self.participants {
-            participant.selectedCard = ticket?.ticketVotes.first(where: { $0.user.id == participant.id })?.selectedCard
-        }
-    }
+
 }
