+++ mamba/mamba/Planning/Shared/Views/PlanningFinishedVotingStateGraphCardView.swift
@@ -19,8 +19,9 @@ struct PlanningFinishedVotingStateGraphCardView: View {
                 .multilineTextAlignment(.center)
                 .padding(leading: 20, top: 15, trailing: 20)
             
-            HorizontalCombinedBarGraphView(barGraphEntries: self.barGraphEntries)
-            .padding(leading: 20, top: 15, bottom: 20, trailing: 20)
+            HorizontalCombinedBarGraphView(barGraphEntries: self.barGraphEntries, barWidth: UIScreen.main.bounds.width - 70)
+                .padding(leading: 20, top: 15, bottom: 20, trailing: 20)
+                .frame(maxWidth: .infinity)
         }
     }
 }
@@ -28,8 +29,9 @@ struct PlanningFinishedVotingStateGraphCardView: View {
 struct PlanningFinishedVotingStateGraphCardView_Previews: PreviewProvider {
     static var previews: some View {
         PlanningFinishedVotingStateGraphCardView(barGraphEntries: [
-            CombinedBarGraphEntry(title: "5", count: 8),
-            CombinedBarGraphEntry(title: "1", count: 2)
+            CombinedBarGraphEntry(title: "5", count: 10),
+            CombinedBarGraphEntry(title: "1", count: 2),
+            CombinedBarGraphEntry(title: "3", count: 1)
         ])
     }
 }
